Add provisioning code to install Droonga Engine on Wily
We cannot install Droonga Engine because on official Ubuntu Wily
image for Vagrant ruby-ffi-yajl package is pre-installed and it
conflicts with ffi-yajl inteded to be installed by gem command.
There is no problem to uninstall ruby-ffi-yajl and related packages
because this is tutorial and this machine is not used in production
environment.